third_party/blink/web_tests/virtual/omt-worker-fetch/external/wpt/workers/interfaces/WorkerGlobalScope/location/redirect-expected.txt

This is a testharness.js-based test.
FAIL WorkerLocation with redirects: classic dedicated workers assert_equals: expected "http://web-platform.test:8001/workers/interfaces/WorkerGlobalScope/location/post-location-members.js?a" but got "http://web-platform.test:8001/workers/interfaces/WorkerGlobalScope/location/helper-redirect.py?fail"
Harness: the test ran to completion.
